fix: slice - actions , reducer 통합작업 slice에 action과 reducer를 통합했습니다
