+++ src/slice.js
@@ -1,6 +1,17 @@
+import { createSlice } from '@reduxjs/toolkit';
+
 import { equal } from './utils';
 
-const { createSlice } = require('@reduxjs/toolkit');
+import {
+  fetchRegions,
+  fetchCategories,
+  fetchRestaurants,
+  fetchRestaurant,
+  postLogin,
+  postReview,
+} from './services/api';
+
+import { saveItem } from './services/storage';
 
 const initialReviewFields = {
   score: '',
@@ -25,36 +36,36 @@ const { actions, reducer } = createSlice({
       ...initialReviewFields,
     },
   },
-  reducer: {
-    setRegions(state, { payload: { regions } }) {
+  reducers: {
+    setRegions(state, { payload: regions }) {
       return {
         ...state,
         regions,
       };
     },
 
-    setCategories(state, { payload: { categories } }) {
+    setCategories(state, { payload: categories }) {
       return {
         ...state,
         categories,
       };
     },
 
-    setRestaurants(state, { payload: { restaurants } }) {
+    setRestaurants(state, { payload: restaurants }) {
       return {
         ...state,
         restaurants,
       };
     },
 
-    setRestaurant(state, { payload: { restaurant } }) {
+    setRestaurant(state, { payload: restaurant }) {
       return {
         ...state,
         restaurant,
       };
     },
 
-    selectRegion(state, { payload: { regionId } }) {
+    selectRegion(state, { payload: regionId }) {
       const { regions } = state;
       return {
         ...state,
@@ -62,7 +73,7 @@ const { actions, reducer } = createSlice({
       };
     },
 
-    selectCategory(state, { payload: { categoryId } }) {
+    selectCategory(state, { payload: categoryId }) {
       const { categories } = state;
       return {
         ...state,
@@ -80,7 +91,7 @@ const { actions, reducer } = createSlice({
       };
     },
 
-    setAccessToken(state, { payload: { accessToken } }) {
+    setAccessToken(state, { payload: accessToken }) {
       return {
         ...state,
         accessToken,
@@ -113,7 +124,7 @@ const { actions, reducer } = createSlice({
       };
     },
 
-    setReviews(state, { payload: { reviews } }) {
+    setReviews(state, { payload: reviews }) {
       const { restaurant } = state;
 
       return {
@@ -127,4 +138,91 @@ const { actions, reducer } = createSlice({
   },
 });
 
-export { actions, reducer };
+export const {
+  setRegions,
+  setCategories,
+  setRestaurants,
+  setRestaurant,
+  setAccessToken,
+  clearReviewFields,
+  setReviews,
+  selectRegion,
+  logout,
+  selectCategory,
+  changeReviewField,
+  changeLoginField,
+} = actions;
+
+export function loadInitialData() {
+  return async (dispatch) => {
+    const regions = await fetchRegions();
+    dispatch(setRegions(regions));
+
+    const categories = await fetchCategories();
+    dispatch(setCategories(categories));
+  };
+}
+
+export function loadRestaurants() {
+  return async (dispatch, getState) => {
+    const {
+      selectedRegion: region,
+      selectedCategory: category,
+    } = getState();
+
+    if (!region || !category) {
+      return;
+    }
+
+    const restaurants = await fetchRestaurants({
+      regionName: region.name,
+      categoryId: category.id,
+    });
+    dispatch(setRestaurants(restaurants));
+  };
+}
+
+export function loadRestaurant({ restaurantId }) {
+  return async (dispatch) => {
+    dispatch(setRestaurant(null));
+
+    const restaurant = await fetchRestaurant({ restaurantId });
+
+    dispatch(setRestaurant(restaurant));
+  };
+}
+
+export function requestLogin() {
+  return async (dispatch, getState) => {
+    const { loginFields: { email, password } } = getState();
+
+    const accessToken = await postLogin({ email, password });
+
+    saveItem('accessToken', accessToken);
+
+    dispatch(setAccessToken(accessToken));
+  };
+}
+
+export function loadReview({ restaurantId }) {
+  return async (dispatch) => {
+    const restaurant = await fetchRestaurant({ restaurantId });
+
+    dispatch(setReviews(restaurant.reviews));
+  };
+}
+
+export function sendReview({ restaurantId }) {
+  return async (dispatch, getState) => {
+    const { accessToken, reviewFields: { score, description } } = getState();
+
+    await postReview({
+      accessToken, restaurantId, score, description,
+    });
+
+    dispatch(loadReview({ restaurantId }));
+    dispatch(clearReviewFields());
+  };
+}
+
+export default reducer;
